fixed the format for hours. wrote down cutoffs.
git-svn-id: file:///mnt/sdd1/kens/svnroot/trunk/coop@502 f16cc87a-5815-0410-8a9a-9fe0dcc0200c

+++ sql/definition.sql
@@ -449,7 +449,7 @@ create table enhancement_hours (
     parentsid int(32),
     enhancement_project_id int(32) not null,  -- join
     work_date date,
-    hours int(5),
+    hours decimal(4,2),
     primary key (enhancement_hour_id)
 );
